% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{bernoulliWCor}
\alias{bernoulliWCor}
\title{bernoulliWCor}
\usage{
bernoulliWCor(n, W = NULL, r, whichCorr, ...)
}
\description{
Draw bernoulli variable with prob x/r (to induce correlation)
}
