% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{bernoulliWCorParm}
\alias{bernoulliWCorParm}
\title{bernoulliWCorParm}
\usage{
bernoulliWCorParm(W = NULL, rLower = -2, rUpper = 2, nCorrLower = 1,
  nCorrUpper = ncol(W))
}
\description{
Generate parameters for \code{bernoulliCorParm} (smaller r = more correlation)
}
