% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{binomialFracWParm}
\alias{binomialFracWParm}
\title{binomialWParm}
\usage{
binomialFracWParm(pLower = 0.1, pUpper = 0.9, numLower = 2,
  numUpper = 10, pFracLower = 0.05, pFracUpper = 0.95, ...)
}
\description{
Generate parameters for \code{binomial}
}
