% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{binomialWParm}
\alias{binomialWParm}
\title{binomialWParm}
\usage{
binomialWParm(pLower = 0.1, pUpper = 0.9, numLower = 2, numUpper = 20,
  ...)
}
\description{
Generate parameters for \code{binomial}
}
