% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{cubicSplineUniParm}
\alias{cubicSplineUniParm}
\title{cubicSplineUniParm}
\usage{
cubicSplineUniParm(x, mainLower = -0.5, mainUpper = 0.5, splineLower = -1,
  splineUpper = 1, nKnotLower = 1, nKnotUpper = 20, quantLower = 0.05,
  quantUpper = 0.95)
}
\description{
cubicSplineUniParm
}
