% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{dNormAddBiv}
\alias{dNormAddBiv}
\title{dNormAddBiv}
\usage{
dNormAddBiv(x1, x2, coef, mult1, mult2, mu, sig)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{coef}{Multiplier in front of dnorm}

\item{mult1}{Multiplier 1 inside dnorm}

\item{mult2}{Multiplier 2 inside dnorm}

\item{mu}{Mean of dnorm}

\item{sig}{SD of dnorm}
}
\description{
Simulate a normal pdf relationship with two-way interaction
}
