% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{dNormAddBivParm}
\alias{dNormAddBivParm}
\title{dNormAddBivParm}
\usage{
dNormAddBivParm(..., coefLower = -5, coefUpper = 5, mult1Lower = -1,
  mult1Upper = 1, mult2Lower = -1, mult2Upper = 1, muLower = -2,
  muUpper = 2, sigLower = 0.5, sigUpper = 2)
}
\description{
Generate parameters for dNormAddBiv
}
