% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{dNormAddQuad}
\alias{dNormAddQuad}
\title{dNormAddQuad}
\usage{
dNormAddQuad(x1, x2, x3, x4, coef, mult1, mult2, mult3, mult4, mu, sig)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{coef}{Multiplier in front of dnorm}

\item{mult1}{Multiplier 1 inside dnorm}

\item{mult2}{Multiplier 2 inside dnorm}

\item{mult3}{Multiplier 3 inside dnnorm}

\item{mult4}{Multiplier 3 inside dnnorm}

\item{mu}{Mean of dnorm}

\item{sig}{SD of dnorm}
}
\description{
Simulate a normal pdf relationship with two-way interaction
}
