% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{dNormMixUni}
\alias{dNormMixUni}
\title{dNormMixUni}
\usage{
dNormMixUni(x, coef1, coef2, mu1, mu2, sig1, sig2)
}
\arguments{
\item{x}{A vector}

\item{coef1}{Multiplier in front of first pdf}

\item{coef2}{Multiplier in front of second pdf}

\item{mu1}{Mean of first pdf}

\item{mu2}{Mean of second pdf}

\item{sig1}{SD of first pdf}

\item{sig2}{SD of second pdf}
}
\description{
Simulate a mixture of normal pdfs relationship with x
}

