% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{dNormMixUniParm}
\alias{dNormMixUniParm}
\title{dNormMixUniParm}
\usage{
dNormMixUniParm(..., coef1Lower = -4, coef1Upper = 4, coef2Lower = -4,
  coef2Upper = 4, mu1Lower = -10, mu1Upper = 10, mu2Lower = -10,
  mu2Upper = 10, sig1Lower = 0.5, sig1Upper = 4, sig2Lower = 0.5,
  sig2Upper = 4)
}
\description{
Generate parameters for \code{dNormMixUni}
}
