% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{dNormMultBiv}
\alias{dNormMultBiv}
\title{dNormMultBiv}
\usage{
dNormMultBiv(x1, x2, coef, mult, mu, sig)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{coef}{Multiplier in front of dnorm}

\item{mult1}{Multiplier 1 inside dnorm}

\item{mult2}{Multiplier 2 inside dnorm}
}
\description{
Simulate a normal pdf relationship with two-way interaction
}

