% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{dNormMultBivParm}
\alias{dNormMultBivParm}
\title{dNormMultBivParm}
\usage{
dNormMultBivParm(x1, x2, coefLower = -5, coefUpper = 5, multLower = -1,
  multUpper = 1, muLower = -2, muUpper = 2, sigLower = 0.5,
  sigUpper = 2)
}
\description{
Generate parameters for dNormMultBiv
}
