% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{dNormMultJump1Tri}
\alias{dNormMultJump1Tri}
\title{dNormMultJump1Tri}
\usage{
dNormMultJump1Tri(x1, x2, x3, coef, mult, mu, sig, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{coef}{Multiplier in front of dnorm}

\item{mult}{Multiplier 1 inside dnorm}

\item{jumpLoc3}{Random quantile for x3 jump}
}
\description{
Simulate a normal pdf relationship with two-way interaction
}
