% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{dNormMultJump1TriParm}
\alias{dNormMultJump1TriParm}
\title{dNormMultTriParm}
\usage{
dNormMultJump1TriParm(x1, x2, x3, coefLower = -5, coefUpper = 5,
  multLower = -1, multUpper = 1, muLower = -2, muUpper = 2,
  quantLower = 0.1, quantUpper = 0.9, sigLower = 0.5, sigUpper = 2)
}
\description{
Generate parameters for dNormMultTri
}
