% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{dNormMultJumpBiv}
\alias{dNormMultJumpBiv}
\title{dNormMultJumpBiv}
\usage{
dNormMultJumpBiv(x1, x2, coef, mult, jumpLoc2, mu, sig)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{coef}{Multiplier in front of dnorm}

\item{mult}{Multiplier 1 inside dnorm}

\item{jumpLoc2}{Location of random quantile jump}
}
\description{
Simulate a normal pdf relationship with two-way interaction
}
