% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{dNormMultQuad}
\alias{dNormMultQuad}
\title{dNormMultQuad}
\usage{
dNormMultQuad(x1, x2, x3, x4, coef, mult, mu, sig)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{coef}{Multiplier in front of dnorm}

\item{mult}{Multiplier 1 inside dnorm}
}
\description{
Simulate a normal pdf relationship with two-way interaction
}
