% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{dNormMultQuadParm}
\alias{dNormMultQuadParm}
\title{dNormMultQuadParm}
\usage{
dNormMultQuadParm(x1, x2, x3, x4, coefLower = -5, coefUpper = 5,
  multLower = -1, multUpper = 1, muLower = -2, muUpper = 2,
  sigLower = 0.5, sigUpper = 2)
}
\description{
Generate parameters for dNormMultQuad
}
