% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{dNormMultTri}
\alias{dNormMultTri}
\title{dNormMultTri}
\usage{
dNormMultTri(x1, x2, x3, coef, mult, mu, sig)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{coef}{Multiplier in front of dnorm}

\item{mult}{Multiplier 1 inside dnorm}
}
\description{
Simulate a normal pdf relationship with two-way interaction
}
