% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{dNormUni}
\alias{dNormUni}
\title{dNormUni}
\usage{
dNormUni(x, coef, mu, sig)
}
\arguments{
\item{x}{A vector}

\item{coef}{The coefficient in front of the normal pdf}

\item{mu}{The mean of the normal pdf}

\item{sig}{The sd of the normal pdf}
}
\description{
Simulate a normal pdf relationship with x
}

