% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{dNormUniParm}
\alias{dNormUniParm}
\title{dNormUniParm}
\usage{
dNormUniParm(..., coefLower = -4, coefUpper = 4, muLower = -5,
  muUpper = 5, sigLower = 0.25, sigUpper = 3)
}
\description{
Generate parameters for \code{dNormUni}
}

