% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBias.R
\name{findBestSims}
\alias{findBestSims}
\title{findBestSims}
\usage{
findBestSims(result, estName, compEst = TRUE, compEstNames, nSims = 100,
  pctTruth = TRUE, threshBest = 0.05, truthName = "truth", ...)
}
\arguments{
\item{result}{The results data frame}

\item{estName}{The estimator that you want to find good simulations for}

\item{compEst}{Define best by comparing to other estimators or by being better
than a certain threshold in absolute error.}

\item{compEstNames}{The estimators that you want to compare the chosen estimator
to in order to define good (e.g., all other estimators)}

\item{nSims}{The maximum number of simulations to return}

\item{pctTruth}{Base on bias as percent of truth or in units of information}

\item{threshBest}{The threshold that defines a best simulation (in pct of truth or
Z-score depend on value of \code{pctTruth})}
}
\description{
Locate the simulations where a particular estimator is performing the best
}
