% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBias.R
\name{findWorstSims}
\alias{findWorstSims}
\title{findWorstSims}
\usage{
findWorstSims(result, estName, compEstNames, nSims = 1000, pctTruth = TRUE,
  compEst = TRUE, threshWorst = 4, truthName = "truth", ...)
}
\arguments{
\item{result}{The results data frame}

\item{estName}{The estimator that you want to find bad simulations for}

\item{compEstNames}{The estimators that you want to compare the chosen estimator
to in order to define bad (e.g., all other estimators)}

\item{nSims}{The maximum number of simulations to return}

\item{pctTruth}{Base on bias as percent of truth or in units of information}

\item{compEst}{Define best by comparing to other estimators or by being better
than a certain threshold in absolute error.}

\item{threshWorst}{The threshold that defines a best simulation (in pct of truth or
Z-score depend on value of \code{pctTruth})}
}
\description{
Locate the simulations where a particular estimator is performing the worst
}
