% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorDistributions.R
\name{gammaErrParm}
\alias{gammaErrParm}
\title{gammaErrParm
Generate parameters for \code{gammaErr}}
\usage{
gammaErrParm(aLower = 0.1, aUpper = 7.5, bLower = 0.1, bUpper = 7.5)
}
\description{
gammaErrParm
Generate parameters for \code{gammaErr}
}
