% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{gammaPointMassWParm}
\alias{gammaPointMassWParm}
\title{gammaWParm
Generate parameters for \code{gammaDist}}
\usage{
gammaPointMassWParm(aLower = 0.5, aUpper = 2.5, bLower = 0.5,
  bUpper = 2.5, pLower = 0.05, pUpper = 0.95, ...)
}
\description{
gammaWParm
Generate parameters for \code{gammaDist}
}
