% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R, R/errorDistributions.R
\name{gammaW}
\alias{gammaW}
\title{gammaW}
\usage{
gammaW(n, a, b, ...)

gammaW(n, a, b, ...)
}
\description{
Draw from gamma(a,b) distribution

Draw an error from a gamma distribution with particular range
}

