% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R, R/errorDistributions.R
\name{gammaWParm}
\alias{gammaWParm}
\title{gammaWParm
Generate parameters for \code{gammaDist}}
\usage{
gammaWParm(aLower = 0.5, aUpper = 4.5, bLower = 0.5, bUpper = 4.5)

gammaWParm(aLower = 0.5, aUpper = 4.5, bLower = 0.5, bUpper = 4.5)
}
\description{
gammaWParm
Generate parameters for \code{gammaDist}

gammaErrParm
Generate parameters for \code{gammaErr}
}

