% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{jumpBivParm}
\alias{jumpBivParm}
\title{jumpBivParm}
\usage{
jumpBivParm(x1, x2, njumpLower = 1, njumpUpper = 5, jumpLower = -2,
  jumpUpper = 2, jumpLoc1 = function(x, njump) {     quantile(x, seq(0, 1,
  length = njump + 1)) }, jumpLoc2 = function(x, njump) {     quantile(x,
  seq(0, 1, length = njump + 1)) })
}
\description{
Generate parameters for \code{jumpBiv}. See source code to see what it does.
}

