% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{jumpQuadParm}
\alias{jumpQuadParm}
\title{jumpQuadParm}
\usage{
jumpQuadParm(x1, x2, x3, x4, njumpLower = 1, njumpUpper = 5,
  jumpLower = -2, jumpUpper = 2, jumpLoc1 = function(x, njump) {    
  quantile(x, seq(0, 1, length = njump + 1)) }, jumpLoc2 = function(x, njump)
  {     quantile(x, seq(0, 1, length = njump + 1)) }, jumpLoc3 = function(x,
  njump) {     quantile(x, seq(0, 1, length = njump + 1)) },
  jumpLoc4 = function(x, njump) {     quantile(x, seq(0, 1, length = njump +
  1)) }, ...)
}
\description{
Generate parameters for \code{jumpQuad}. See source code and \code{?jumpUni}.
}
