% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{jumpTri}
\alias{jumpTri}
\title{jumpTri}
\usage{
jumpTri(x1, x2, x3, njump, jumpLoc1, jumpLoc2, jumpLoc3, jumpVal)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{njump}{The number of jumps the function takes}

\item{jumpLoc1}{A vector of the locations of those jumps for x1}

\item{jumpLoc2}{A vector of the locations of those jumps for x2}

\item{jumpLoc3}{A vector of the locations of those jumps for x3}

\item{jumpVal}{A vector of length njump indicating the value of the function at each jump}
}
\value{
The value of the jump function.
}
\description{
Simulate a jump function trivariate relationship
}
