% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{jumpUni}
\alias{jumpUni}
\title{jumpUni}
\usage{
jumpUni(x, njump, jumpLoc, jumpVal)
}
\arguments{
\item{x}{A vector}

\item{njump}{The number of jumps the function takes}

\item{jumpLoc}{A vector of the locations of those jumps}

\item{jumpVal}{A vector of length njump indicating the value of the function at each jump}
}
\value{
The value of the jump function.
}
\description{
Simulate a jump function univariate relationship
}
