% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{jumpUniParm}
\alias{jumpUniParm}
\title{jumpUniParm}
\usage{
jumpUniParm(x, njumpLower = 1, njumpUpper = 8, jumpLower = -4,
  jumpUpper = 4, jumpLoc = function(x, njump) {     quantile(x, seq(0, 1,
  length = njump + 1)) })
}
\description{
Generate parameters for \code{jumpUni}. Here jumpLoc is a function that takes x and njump as input and outputs
the locations of jumps. By contrast, when jumpLoc is fed into jumpUni, it is just a vector
of numbers. See source code for how exactly it works.
}
