% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{linBiv}
\alias{linBiv}
\title{linBiv}
\usage{
linBiv(x1, x2, coef)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{coef}{A numeric (supplied by \code{bivUniParm})}
}
\value{
\code{coef*x1*x2}
}
\description{
Simulate a bivariate univariate relationship
}

