% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{linJump1Quad}
\alias{linJump1Quad}
\title{linJump1Quad}
\usage{
linJump1Quad(x1, x2, x3, x4, coef, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{coef}{A numeric (supplied by \code{bivUniParm})}

\item{jumpLoc3}{A numeric quantile for x3}
}
\value{
\code{coef*x1*x2*x3}
}
\description{
Simulate a trivariate univariate relationship
}
