% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{linJump1QuadParm}
\alias{linJump1QuadParm}
\title{linQuadParm}
\usage{
linJump1QuadParm(coefLower = -10, coefUpper = 10, quantLower = 0.1,
  quantUpper = 0.9, x1, x2, x3, x4)
}
\description{
Generate parameters for \code{linQuad}.
}
