% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{linJump1Tri}
\alias{linJump1Tri}
\title{linJump1Tri}
\usage{
linJump1Tri(x1, x2, x3, coef, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{coef}{A numeric (supplied by \code{bivUniParm})}

\item{quant3}{A numeric quantile for x3}
}
\value{
\code{coef*x1*x2*x3}
}
\description{
Simulate a trivariate univariate relationship
}
