% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{linJump1TriParm}
\alias{linJump1TriParm}
\title{linTriParm}
\usage{
linJump1TriParm(x1, x2, x3, coefLower = -10, coefUpper = 10,
  quantLower = 0.1, quantUpper = 0.9)
}
\description{
Generate parameters for \code{linTri}.
}
