% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{linJump2Quad}
\alias{linJump2Quad}
\title{linJump2Quad}
\usage{
linJump2Quad(x1, x2, x3, x4, coef, jumpLoc2, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{coef}{A numeric (supplied by \code{bivUniParm})}

\item{jumpLoc2}{Random quantile for x2 jump}

\item{jumpLoc3}{Random quantile for x3 jump}
}
\value{
\code{coef*x1*x2*x3}
}
\description{
Simulate a trivariate univariate relationship
}
