% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{linJump2QuadParm}
\alias{linJump2QuadParm}
\title{linJump2QuadParm}
\usage{
linJump2QuadParm(coefLower = -10, coefUpper = 10, quantLower = 0.1,
  quantUpper = 0.9, x1, x2, x3, x4)
}
\description{
Generate parameters for \code{linQuad}.
}
