% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{linJump2Tri}
\alias{linJump2Tri}
\title{linJump2Tri}
\usage{
linJump2Tri(x1, x2, x3, coef, jumpLoc2, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{coef}{A numeric (supplied by \code{bivUniParm})}

\item{jumpLoc2}{Random quantile value for x2}

\item{jumpLoc3}{Random quantile value for x2}
}
\value{
\code{coef*x1*x2*x3}
}
\description{
Simulate a trivariate univariate relationship
}
