% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{linJumpBiv}
\alias{linJumpBiv}
\title{linJumpBiv}
\usage{
linJumpBiv(x1, x2, coef, jumpLoc)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{coef}{A numeric (supplied by \code{linBivParm})}

\item{jumpLoc}{Location of randomly chosen quantile}
}
\value{
\code{coef*x1*x2}
}
\description{
Simulate a bivariate relationship with jump at random quantile of 
one variable.
}
