% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{linJumpBivParm}
\alias{linJumpBivParm}
\title{linJumpBivParm}
\usage{
linJumpBivParm(coefLower = -4, coefUpper = 4, quantLower = 0.1,
  quantUpper = 0.9, x1, x2)
}
\description{
Generate coefficient for \code{linBiv}
}
