% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{linQuad}
\alias{linQuad}
\title{linQuad}
\usage{
linQuad(x1, x2, x3, x4, coef)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{coef}{A numeric (supplied by \code{bivUniParm})}
}
\value{
\code{coef*x1*x2*x3}
}
\description{
Simulate a trivariate univariate relationship
}
