% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{linSplineBivParm}
\alias{linSplineBivParm}
\title{linSplineBivParm}
\usage{
linSplineBivParm(x1, x2, slopeLower = -3, slopeUpper = 3, nKnotLower = 1,
  nKnotUpper = 5, quantLower = 0.05, quantUpper = 0.95)
}
\description{
linSplineBivParm
}
