% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{linSplineUniParm}
\alias{linSplineUniParm}
\title{linSplineUniParm}
\usage{
linSplineUniParm(x, slopeLower = -10, slopeUpper = 10, nKnotLower = 1,
  nKnotUpper = 5, quantLower = 0.05, quantUpper = 0.95)
}
\description{
linSplineUniParm
}
