% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{linUni}
\alias{linUni}
\title{linUni}
\usage{
linUni(x, coef)
}
\arguments{
\item{x}{A vector}

\item{coef}{A numeric (supplied by \code{linUniParm})}
}
\value{
\code{coef*x}
}
\description{
Simulate a linear univariate relationship
}
