% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorDistributions.R
\name{normalErrAW}
\alias{normalErrAW}
\title{normalErrAW}
\usage{
normalErrAW(n, sd, co, AW)
}
\description{
Draw an error from a normal distribution with a particular variance that depends on one W
}
