% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorDistributions.R
\name{normalErrAWParm}
\alias{normalErrAWParm}
\title{normalErrAWParm}
\usage{
normalErrAWParm(AW, sdLower = 0.5, sdUpper = 7, nDepLower = 0.5,
  nDepUpper = 3.5)
}
\description{
Generate parameters for \code{normalErrUniW}
}
