% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{normalPointMassWParm}
\alias{normalPointMassWParm}
\title{normalPointMassWParm}
\usage{
normalPointMassWParm(meanLower = -2, meanUpper = 2, sdLower = 0.25,
  sdUpper = 2, pLower = 0.05, pUpper = 0.95, ...)
}
\description{
Generate parameters for \code{normal}
}
