% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{normalWCor}
\alias{normalWCor}
\title{normalWCor}
\usage{
normalWCor(n, W = NULL, sd, whichCorr, ...)
}
\description{
Draw normal variate with mean equal to sum of the selected variables
that it is correlated with
}
