% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{normalWCorParm}
\alias{normalWCorParm}
\title{normalWCorParm}
\usage{
normalWCorParm(W = NULL, sdLower = 0.25, sdUpper = 4, nCorrLower = 1,
  nCorrUpper = ncol(W))
}
\description{
Generate parameters for \code{normalCor} (smaller SD = more correlation)
}
