% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{pLogisAddBiv}
\alias{pLogisAddBiv}
\alias{pLogisAddBiv}
\title{pLogisAddBiv}
\usage{
pLogisAddBiv(x1, x2, coefLower = -4, coefUpper = 4, mult1Lower = -1,
  mult1Upper = 1, mult2Lower = -1, mult2Upper = 1, locLower = -2,
  locUpper = 2, scaleLower = 0.25, scaleUpper = 2)

pLogisAddBiv(x1, x2, coefLower = -4, coefUpper = 4, mult1Lower = -1,
  mult1Upper = 1, mult2Lower = -1, mult2Upper = 1, locLower = -2,
  locUpper = 2, scaleLower = 0.25, scaleUpper = 2)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{coef}{Multiplier in front of plogis}

\item{mult1}{Multiplier 1 inside plogis}

\item{mult2}{Multiplier 2 inside plogis}

\item{loc}{The location parameter for plogis}

\item{scale}{The scale parameter}
}
\description{
pLogisAddBiv

Generate parameters for \code{pLogisAddBiv}
}
