% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{pLogisAddQuad}
\alias{pLogisAddQuad}
\title{pLogisAddQuad}
\usage{
pLogisAddQuad(x1, x2, x3, x4, coef, mult1, mult2, mult3, mult4, loc, scale)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{coef}{Multiplier in front of plogis}

\item{mult1}{Multiplier 1 inside plogis}

\item{mult2}{Multiplier 2 inside plogis}

\item{mult3}{Multiplier 3 inside plogis}

\item{mult4}{Multiplier 4 inside plogis}

\item{loc}{The location parameter for plogis}

\item{scale}{The scale parameter}
}
\description{
pLogisAddQuad
}
