% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{pLogisAddQuadParm}
\alias{pLogisAddQuadParm}
\title{pLogisAddQuadParm}
\usage{
pLogisAddQuadParm(x1, x2, x3, x4, coefLower = -4, coefUpper = 4,
  multLower = -1, multUpper = 1, locLower = -2, locUpper = 2,
  scaleLower = 0.25, scaleUpper = 2)
}
\description{
Generate parameters for \code{pLogisAddQuad}
}
