% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{pLogisAddTri}
\alias{pLogisAddTri}
\title{pLogisAddTri}
\usage{
pLogisAddTri(x1, x2, x3, coef, mult1, mult2, mult3, loc, scale)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{coef}{Multiplier in front of plogis}

\item{mult1}{Multiplier 1 inside plogis}

\item{mult2}{Multiplier 2 inside plogis}

\item{mult3}{Multiplier 3 inside plogis}

\item{loc}{The location parameter for plogis}

\item{scale}{The scale parameter}
}
\description{
pLogisAddTri
}
