% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{pLogisAddTriParm}
\alias{pLogisAddTriParm}
\title{pLogisAddTriParm}
\usage{
pLogisAddTriParm(x1, x2, x3, coefLower = -4, coefUpper = 4,
  multLower = -1, multUpper = 1, locLower = -2, locUpper = 2,
  scaleLower = 0.25, scaleUpper = 2)
}
\description{
Generate parameters for \code{pLogisAddTri}
}
