% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{pLogisUni}
\alias{pLogisUni}
\title{pLogisUni}
\usage{
pLogisUni(x, coef, mult, loc, scale)
}
\arguments{
\item{x}{A vector}

\item{coef}{The coefficient in front of the normal pdf}

\item{mult}{The multiplier inside the expit function}
}
\description{
Simulate an expit relationship with x
}
