% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{pLogisUniParm}
\alias{pLogisUniParm}
\title{pLogisUniParm}
\usage{
pLogisUniParm(..., coefLower = -4, coefUpper = 4, multLower = -4,
  multUpper = 4, locLower = -4, locUpper = 4, scaleLower = 0.25,
  scaleUpper = 4)
}
\description{
Parameters for \code{pLogisUni}
}
