% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBias.R
\name{plotBias}
\alias{plotBias}
\title{plotBias}
\usage{
plotBias(result, estNames, strat, truthName = "truth",
  pointOptions = list(pch = 1:length(estNames)), estLabels = estNames,
  legendOptions = list(x = "topleft", legend = estLabels), ylim = NULL,
  pctTruth = TRUE, summary = list(fn = "quantile", options = list(probs =
  0.5)), ...)
}
\arguments{
\item{result}{The result data frame}

\item{estNames}{The names of the estimators to plot}

\item{strat}{The name of the variable on which to stratify}

\item{truthName}{The column name containing the true ATE}

\item{pointOptions}{A list of arguments to be passed to points}

\item{estLabels}{Vector of names to label estimators in legend (default is to use estNames)}

\item{legendOptions}{A list of arguments passed to legend. pointOptions are added
automatically}

\item{ylim}{y-axis limits for the plot. If null, looks for range and adds 5% to it}

\item{pctTruth}{Report error as absolute percent of true ATE? If \code{FALSE}, error is reported as 
absolute units of information (i.e., inverse of variance of EIF)}

\item{summary}{A list with two entries named "fn" and "options". This is the summary
function used to summarize percent difference. Defaults to median (i.e., \code{function = quantile},
\code{options = list(probs = 0.5)}). Assumes that the function will play nicely with the unnamed
argument passed in}

\item{...}{Additional arguments passed to plot}
}
\description{
Function to plot the median percent bias for a given set of estimators
stratified by a given set of covariates
}
