% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBias.R
\name{plotDiff}
\alias{plotDiff}
\title{plotDiff}
\usage{
plotDiff(result, estNames, estLabels = estNames, truthName = "truth",
  ylim = NULL, xlim = NULL, ...)
}
\arguments{
\item{result}{Result data frame}

\item{estNames}{Column names of results data frame that are the two estimators to be compared}

\item{estLabels}{Vector of names to label estimators in legend (default is to use estNames)}

\item{truthName}{Column name of results data frame that is the true ATE}

\item{ylim}{y-axis limits for the plot. If null, looks for range and adds 5% to it}

\item{...}{Other arguments passed to plot}
}
\description{
Plot the difference between the scaled version of two estimators. The estimator 
names are entered in \code{estNames} and the difference between absolute value of
Z-scores for estNames[1] - estNames[2] is plotted.
}
