% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBias.R
\name{plotSamplingDist}
\alias{plotSamplingDist}
\title{plotSamplingDist}
\usage{
plotSamplingDist(result, estNames, estLabels = estNames,
  legendOptions = list(x = "topleft", legend = c(estLabels, "N(0,1)")),
  incAsymDist = TRUE, densityOptions = list(n = 2^12),
  truthName = "truth", ylim = NULL, xlim = NULL, ...)
}
\arguments{
\item{result}{Result data frame}

\item{estNames}{Column name(s) of results data frame that are estimators}

\item{estLabels}{Vector of names to label estimators in legend (default is to use estNames)}

\item{legendOptions}{A list of arguments passed to legend.}

\item{incAsymDist}{Whether to also include the asymptotic distribution}

\item{densityOptions}{A list of options passed to \code{density}}

\item{truthName}{Column name of results data frame that is the true ATE}

\item{ylim}{y-axis limits for the plot. If null, looks for range and adds 5% to it}

\item{...}{Other arguments passed to plot}
}
\description{
Plot the centered, scaled (by sqrt(var(EIF)/n)) sampling distribution 
of a given set of estimators.
}
