% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{polyBivParm}
\alias{polyBivParm}
\title{polyBivParm}
\usage{
polyBivParm(npolyLower = 1, npolyUpper = 1, coefLower = -0.25,
  coefUpper = 0.25, deg1Lower = 1, deg1Upper = 3, deg2Lower = 1,
  deg2Upper = 3, ...)
}
\description{
Generate parameters for \code{polyBiv}.
}
