% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{polyJump1QuadParm}
\alias{polyJump1QuadParm}
\title{polyQuadParm}
\usage{
polyJump1QuadParm(npolyLower = 1, npolyUpper = 1, coefLower = -0.1,
  coefUpper = 0.1, deg1Lower = 1, deg1Upper = 2, deg2Lower = 1,
  deg2Upper = 2, deg4Lower = 1, deg4Upper = 2, quantLower = 0.1,
  quantUpper = 0.9, x1, x2, x3, x4)
}
\description{
Generate parameters for \code{polyQuad}. See source code.
}
