% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{polyJump1Tri}
\alias{polyJump1Tri}
\title{polyJump1Tri
Simulate a polynomial trivariate relationship}
\usage{
polyJump1Tri(x1, x2, x3, npoly, coef, deg1, deg2, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{npoly}{A numeric indicating the number of polynomial terms to simulate}

\item{coef}{A vector indicating the coefficient for each polynomial term}

\item{deg1}{A vector indicating the degree of the first polynomial term}

\item{deg2}{A vector indicating the degree of the second polynomial term}

\item{jumpLoc3}{Location of random quantile jump for x3}

\item{coef}{A numeric coefficient}
}
\value{
Trivariate polynomial (see source code for details)
}
\description{
polyJump1Tri
Simulate a polynomial trivariate relationship
}
