% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{polyJump1TriParm}
\alias{polyJump1TriParm}
\title{polyTriParm}
\usage{
polyJump1TriParm(npolyLower = 1, npolyUpper = 1, coefLower = -0.1,
  coefUpper = 0.1, deg1Lower = 1, deg1Upper = 2, deg2Lower = 1,
  deg2Upper = 2, quantLower = 0.1, quantUpper = 0.9, x1, x2, x3)
}
\description{
Generate parameters for \code{polyTri}. See source code.
}
