% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{polyJump2Tri}
\alias{polyJump2Tri}
\title{polyJump2Tri
Simulate a polynomial trivariate relationship}
\usage{
polyJump2Tri(x1, x2, x3, npoly, coef, deg1, jumpLoc2, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{npoly}{A numeric indicating the number of polynomial terms to simulate}

\item{coef}{A vector indicating the coefficient for each polynomial term}

\item{deg1}{A vector indicating the degree of the first polynomial term}

\item{jumpLoc2}{Random quantile jump of x2}

\item{jumpLoc3}{Random quantile jump of x2}

\item{coef}{A numeric coefficient}
}
\value{
Trivariate polynomial (see source code for details)
}
\description{
polyJump2Tri
Simulate a polynomial trivariate relationship
}
