% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{polyJumpBiv}
\alias{polyJumpBiv}
\title{polyJumpBiv
Simulate a polynomial bivariate relationship}
\usage{
polyJumpBiv(x1, x2, npoly, coef, deg1, jumpLoc)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{npoly}{A numeric indicating the number of polynomial terms to simulate}

\item{coef}{A vector indicating the coefficient for each polynomial term}

\item{deg1}{A vector indicating the degree of the first polynomial term}

\item{jumpLoc}{Location of randomly chosen quantile}

\item{coef}{A numeric coefficient}
}
\value{
\code{coef[1]*x1^deg1[1]*x2^deg2[2] + ... + coef[npoly]*x1^deg1[npoly]*x2^deg2[2]}
}
\description{
polyJumpBiv
Simulate a polynomial bivariate relationship
}
