% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{polyJumpBivParm}
\alias{polyJumpBivParm}
\title{polyJumpBivParm}
\usage{
polyJumpBivParm(npolyLower = 1, npolyUpper = 2, coefLower = -0.25,
  coefUpper = 0.25, deg1Lower = 1, deg1Upper = 3, quantLower = 0.1,
  quantUpper = 0.9, x1, x2)
}
\description{
Generate parameters for \code{polyBiv}.
}
