% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{polyQuad}
\alias{polyQuad}
\title{polyQuad
Simulate a polynomial trivariate relationship}
\usage{
polyQuad(x1, x2, x3, x4, npoly, coef, deg1, deg2, deg3, deg4)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{npoly}{A numeric indicating the number of polynomial terms to simulate}

\item{coef}{A vector indicating the coefficient for each polynomial term}

\item{deg1}{A vector indicating the degree of the first polynomial term}

\item{deg2}{A vector indicating the degree of the second polynomial term}

\item{deg3}{A vector indicating the degree of the second polynomial term}

\item{coef}{A numeric coefficient}
}
\value{
Quadvariate polynomial (see source code for details)
}
\description{
polyQuad
Simulate a polynomial trivariate relationship
}
