% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{polyQuadParm}
\alias{polyQuadParm}
\title{polyQuadParm}
\usage{
polyQuadParm(npolyLower = 1, npolyUpper = 1, coefLower = -0.1,
  coefUpper = 0.1, deg1Lower = 1, deg1Upper = 2, deg2Lower = 1,
  deg2Upper = 2, deg3Lower = 1, deg3Upper = 2, deg4Lower = 1,
  deg4Upper = 2, x1, x2, x3, x4)
}
\description{
Generate parameters for \code{polyQuad}. See source code.
}
