% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{polyTri}
\alias{polyTri}
\title{polyTri
Simulate a polynomial trivariate relationship}
\usage{
polyTri(x1, x2, x3, coef, deg1, deg2, deg3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{coef}{A vector indicating the coefficient for each polynomial term}

\item{deg1}{A vector indicating the degree of the first polynomial term}

\item{deg2}{A vector indicating the degree of the second polynomial term}

\item{deg3}{A vector indicating the degree of the second polynomial term}

\item{npoly}{A numeric indicating the number of polynomial terms to simulate}

\item{coef}{A numeric coefficient}
}
\value{
Trivariate polynomial (see source code for details)
}
\description{
polyTri
Simulate a polynomial trivariate relationship
}

