% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{polyTriParm}
\alias{polyTriParm}
\title{polyTriParm}
\usage{
polyTriParm(npolyLower = 1, npolyUpper = 1, coefLower = -0.1,
  coefUpper = 0.1, deg1Lower = 1, deg1Upper = 2, deg2Lower = 1,
  deg2Upper = 2, deg3Lower = 1, deg3Upper = 2, ...)
}
\description{
Generate parameters for \code{polyTri}. See source code.
}

