% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{polyUni}
\alias{polyUni}
\title{polyUni}
\usage{
polyUni(x, npoly, coef, deg)
}
\arguments{
\item{x}{A vector}

\item{npoly}{A numeric indicating the number of polynomial terms to simulate}

\item{coef}{A vector indicating the coefficient for each polynomial term}

\item{deg}{A vector indicating the degree of each polynomial term}

\item{coef}{A numeric (supplied by \code{linUniParm})}
}
\value{
\code{coef[1]*x^deg[1] + ... + coef[npoly]*x^deg[npoly]}
}
\description{
Simulate a polynomial univariate relationship
}

