% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{qGammaUni}
\alias{qGammaUni}
\title{qGammaUni}
\usage{
qGammaUni(x, coef, a, b, ...)
}
\arguments{
\item{x}{A vector}

\item{coef}{The coefficient in front of the gamma cdf}

\item{a}{The shape parameter}

\item{b}{The rate parameter}
}
\description{
Simulate a gamma cdf relationship with expit(x)
}

