% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{qGammaUniParm}
\alias{qGammaUniParm}
\title{qGammaUniParm}
\usage{
qGammaUniParm(coefLower = -4, coefUpper = 4, aLower = 0.25, aUpper = 7,
  bLower = 0.25, bUpper = 7, ...)
}
\description{
Generate parameters for \code{qGammaUni}.
}
