% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{sinBiv}
\alias{sinBiv}
\title{sinBiv}
\usage{
sinBiv(x1, x2, amp, p)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{amp}{The amplitude of the sin function}

\item{p}{The periodicity of the sin function}
}
\value{
\code{amp*sin(p*x)}
}
\description{
Simulate a sinusoidal bivariate relationship
}

