% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{sinJump1QuadParm}
\alias{sinJump1QuadParm}
\title{sinJump1QuadParm}
\usage{
sinJump1QuadParm(pLower = -1, pUpper = 1, ampLower = -1, ampUpper = 1,
  quantLower = 0.1, quantUpper = 0.9, x1, x2, x3, x4)
}
\description{
Generate parameters for \code{sinQuad}. See source code.
}
