% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{sinJump1Tri}
\alias{sinJump1Tri}
\title{sinJump1Tri}
\usage{
sinJump1Tri(x1, x2, x3, p, amp, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{p}{The periodicity of the sin function}

\item{amp}{The amplitude of the sin function}

\item{jumpLoc3}{Random jump location for x3}
}
\value{
\code{amp*sin(p*x1*x2*x3)}
}
\description{
Simulate a sinusoidal trivariate relationship
}
