% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{sinJump1TriParm}
\alias{sinJump1TriParm}
\title{sinJump1TriParm}
\usage{
sinJump1TriParm(pLower = -1, pUpper = 1, ampLower = -1, ampUpper = 1,
  quantLower = 0.1, quantUpper = 0.9, x1, x2, x3)
}
\description{
Generate parameters for \code{sinTri}. See source code.
}
