% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{sinJump2Quad}
\alias{sinJump2Quad}
\title{sinJump2Quad}
\usage{
sinJump2Quad(x1, x2, x3, x4, p, amp, jumpLoc2, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{p}{The periodicity of the sin function}

\item{amp}{The amplitude of the sin function}

\item{jumpLoc2}{Random quantile for x2 jump.}

\item{jumpLoc3}{Random quantile for x2 jump.}
}
\value{
\code{amp*sin(p*x1*x2*x3)}
}
\description{
Simulate a sinusoidal trivariate relationship
}
