% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{sinJump2Tri}
\alias{sinJump2Tri}
\title{sinJump2Tri}
\usage{
sinJump2Tri(x1, x2, x3, p, amp, jumpLoc2, jumpLoc3)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{p}{The periodicity of the sin function}

\item{amp}{The amplitude of the sin function}

\item{jumpLoc2}{Random quantile jump for x2}

\item{jumpLoc3}{Random quantile jump for x2}
}
\value{
\code{amp*sin(p*x1*x2*x3)}
}
\description{
Simulate a sinusoidal trivariate relationship
}
