% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{sinJumpBiv}
\alias{sinJumpBiv}
\title{sinJumpBiv}
\usage{
sinJumpBiv(x1, x2, amp, p, jumpLoc)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{amp}{The amplitude of the sin function}

\item{p}{The periodicity of the sin function}

\item{jumpLoc}{Location of randomly chosen quantile}
}
\value{
\code{amp*sin(p*x)}
}
\description{
Simulate a sinusoidal bivariate relationship
}
