% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateFunctions.R
\name{sinJumpBivParm}
\alias{sinJumpBivParm}
\title{sinJumpBivParm}
\usage{
sinJumpBivParm(pLower = -1, pUpper = 1, ampLower = -1, ampUpper = 1,
  quantLower = 0.1, quantUpper = 0.9, x1, x2)
}
\description{
Generate parameters for \code{sinBiv}
}
