% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{sinQuad}
\alias{sinQuad}
\title{sinQuad}
\usage{
sinQuad(x1, x2, x3, x4, p, amp)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{x4}{A vector}

\item{p}{The periodicity of the sin function}

\item{amp}{The amplitude of the sin function}
}
\value{
\code{amp*sin(p*x1*x2*x3)}
}
\description{
Simulate a sinusoidal trivariate relationship
}
