% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadFunctions.R
\name{sinQuadParm}
\alias{sinQuadParm}
\title{sinQuadParm}
\usage{
sinQuadParm(pLower = -1, pUpper = 1, ampLower = -1, ampUpper = 1, x1,
  x2, x3, x4)
}
\description{
Generate parameters for \code{sinQuad}. See source code.
}
