% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{sinTri}
\alias{sinTri}
\title{sinTri}
\usage{
sinTri(x1, x2, x3, p, amp)
}
\arguments{
\item{x1}{A vector}

\item{x2}{A vector}

\item{x3}{A vector}

\item{p}{The periodicity of the sin function}

\item{amp}{The amplitude of the sin function}
}
\value{
\code{amp*sin(p*x1*x2*x3)}
}
\description{
Simulate a sinusoidal trivariate relationship
}
