% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{sinTriParm}
\alias{sinTriParm}
\title{sinTriParm}
\usage{
sinTriParm(pLower = -1, pUpper = 1, ampLower = -1, ampUpper = 1, ...)
}
\description{
Generate parameters for \code{sinTri}. See source code.
}

