% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{sinUni}
\alias{sinUni}
\title{sinUni}
\usage{
sinUni(x, p, amp)
}
\arguments{
\item{x}{A vector}

\item{p}{The periodicity of the sin function}

\item{amp}{The amplitude of the sin function}
}
\value{
\code{amp*sin(p*x)}
 @export
}
\description{
Simulate a sinusoidal univariate relationship
}

