% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{sinUniParm}
\alias{sinUniParm}
\title{sinUniParm}
\usage{
sinUniParm(pLower = -4, pUpper = 4, ampLower = -4, ampUpper = 4, ...)
}
\description{
Draw random periodicity and amplitude for sinusoidal relationship univariate relationship.
See source code for distributions.
}
