% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBias.R
\name{summarizeBias}
\alias{summarizeBias}
\title{summarizeBias}
\usage{
summarizeBias(result, estNames, truthName = "truth", ...)
}
\arguments{
\item{result}{Result data frame}

\item{estNames}{Column name(s) of results data frame that are estimators}

\item{truthName}{Column name of results data frame that is the true ATE}

\item{...}{Other arguments}
}
\value{
A vector with the ordered bias of the estimators
}
\description{
Function to summarize bias of estimators across
different simulations.
}
