% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBias.R
\name{summarizeCoverage}
\alias{summarizeCoverage}
\title{summarizeCoverage}
\usage{
summarizeCoverage(result, cilNames, ciuNames, truthName = "truth", ...)
}
\arguments{
\item{result}{Result data frame}

\item{cilNames}{Column name(s) of results data frame that are lower confidence bounds}

\item{ciuNames}{Column name(s) of results data frame that are upper confidence bounds}

\item{truthName}{Column name of results data frame that is the true ATE}

\item{...}{Other arguments}
}
\value{
A vector with the coverage of the estimators
}
\description{
Function to summarize coverage of confidence intervals across all
simulations
}
