% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRandomData-method-summary.R
\name{summary.makeRandomData}
\alias{summary.makeRandomData}
\title{summary.makeRandomData}
\usage{
\method{summary}{makeRandomData}(object, N = 1e+06, seed = 1234, ...)
}
\description{
A method to summarize output of makeRandomData
}
