% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorDistributions.R
\name{uniformErrAWParm}
\alias{uniformErrAWParm}
\title{uniformErrAWParm}
\usage{
uniformErrAWParm(AW, rangeLower = 1, rangeUpper = 25, nDepLower = 0.5,
  nDepUpper = 3.5)
}
\description{
Generate parameters for \code{uniformErr}
}
