% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{uniformWCorParm}
\alias{uniformWCorParm}
\title{uniformWParm
Generate parameters for \code{uniformParm} (smaller lower = more correlation)}
\usage{
uniformWCorParm(W = NULL, lowerLower = -5, lowerUpper = 0.25,
  rangeLower = 0.25, rangeUpper = 10, nCorrLower = 1,
  nCorrUpper = ncol(W))
}
\description{
uniformWParm
Generate parameters for \code{uniformParm} (smaller lower = more correlation)
}
