% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{uniformWParm}
\alias{uniformWParm}
\title{uniformWParm}
\usage{
uniformWParm(lowerLower = -2, lowerUpper = 0, rangeLower = 0.25,
  rangeUpper = 5)

uniformWParm(lowerLower = -2, lowerUpper = 0, rangeLower = 0.25,
  rangeUpper = 5)
}
\description{
Generate parameters for \code{uniform}

uniformWParm
Generate parameters for \code{uniformParm} (smaller lower = more correlation)
}

