% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{bernoulliWCorParm}
\alias{bernoulliWCorParm}
\title{bernoulliWCorParm}
\usage{
bernoulliWCorParm(rLower = 0.1, rUpper = 10)
}
\description{
Generate parameters for \code{bernoulliCorParm} (smaller r = more correlation)
}

