% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{dNormMixUniParm}
\alias{dNormMixUniParm}
\title{dNormMixUniParm}
\usage{
dNormMixUniParm(..., coef1Lower = -2, coef1Upper = 2, coef2Lower = -2,
  coef2Upper = 2, mu1Lower = -5, mu1Upper = 5, mu2Lower = -5,
  mu2Upper = 5, sig1Lower = 0.5, sig1Upper = 2, sig2Lower = 0.5,
  sig2Upper = 2)
}
\description{
Generate parameters for \code{dNormMixUni}
}

