% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{gammaWParm}
\alias{gammaWParm}
\title{gammaWParm
Generate parameters for \code{gammaDist}}
\usage{
gammaWParm(aLower = 0.5, aUpper = 2.5, bLower = 0.5, bUpper = 2.5, ...)
}
\description{
gammaWParm
Generate parameters for \code{gammaDist}
}
