% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{normalWCorParm}
\alias{normalWCorParm}
\title{normalWCorParm}
\usage{
normalWCorParm(sdLower = 0.25, sdUpper = 2)
}
\description{
Generate parameters for \code{normalCor} (smaller SD = more correlation)
}
