% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{pLogisUniParm}
\alias{pLogisUniParm}
\title{pLogisUniParm}
\usage{
pLogisUniParm(..., coefLower = -2, coefUpper = 2, multLower = -2,
  multUpper = 2, locLower = -2, locUpper = 2, scaleLower = 0.25,
  scaleUpper = 2)
}
\description{
Parameters for \code{pLogisUni}
}
