% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRandomData-method-plot.R
\name{plot.makeRandomData}
\alias{plot.makeRandomData}
\title{plot.makeRandomData}
\usage{
\method{plot}{makeRandomData}(object, g0 = TRUE, Q0 = TRUE, ask = FALSE)
}
\description{
Plot the results of makeRandomData
}

