% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivariateFunctions.R
\name{polyTriParm}
\alias{polyTriParm}
\title{polyBivParm}
\usage{
polyTriParm(npolyLower = 1, npolyUpper = 2, coefLower = -0.25,
  coefUpper = 0.25, deg1Lower = 1, deg1Upper = 3, deg2Lower = 1,
  deg2Upper = 3, deg3Lower = 1, deg3Upper = 3, ...)
}
\description{
Generate parameters for \code{polyTri}. See source code.
}

