% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{polyUniParm}
\alias{polyUniParm}
\title{polyUniParm}
\usage{
polyUniParm(npolyLower = 1, npolyUpper = 3, coefLower = -1,
  coefUpper = 1, degLower = 1, degUpper = 3, ...)
}
\description{
Draw random coefficient for polynomial relationship univariate relationship.
See source code for distributions. 
 @export
}
