% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{qGammaUniParm}
\alias{qGammaUniParm}
\title{qGammaUniParm}
\usage{
qGammaUniParm(coefLower = -2, coefUpper = 2, aLower = 0.5, aUpper = 5,
  bLower = 0.5, bUpper = 5, ...)
}
\description{
Generate parameters for \code{qGammaUni}.
}
