% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remakeRandomData.R
\name{remakeRandomData}
\alias{remakeRandomData}
\title{remakeRandomData}
\usage{
remakeRandomData(n, object, setA = NULL, ...)
}
\arguments{
\item{n}{A \code{numeric} sample size}

\item{maxD}{A \code{numeric} indicating maximum number of covariates}

\item{distW.uncor}{A \code{vector} of \code{characters} that are functions in the global environment that 
generate covariates. These functions should have an associated parameter function with \code{"Parm"} added 
to the end of their name. See the format of the functions provided to understand their structure.}

\item{distW.cor}{Same as above, but generates a correlated covariate}

\item{funcG0.uni}{A \code{vector} of \code{characters} that are functions in the global environment that
are used to generate main terms for logit of the propensity score.}

\item{funcG0.biv}{Ditto above, but for bivariate functions.}

\item{funcG0.tri}{Ditto above, but for trivariate functions.}

\item{funcQ0.uni}{Ditto \code{funcG0.uni}, but for the outcome regression}

\item{funcQ0.biv}{Ditto}

\item{funcQ0.tri}{Ditto}

\item{errY}{A \code{vector} of \code{characters} that are functions in the global environment that 
are used to generate error terms for the outcome regression.}

\item{minG0}{The minimum value for the propensity score (default 0.01).}
}
\value{
An object of class \code{"makeRandomData"} with the following entries
\item{W}{A matrix of covariates}
\item{A}{A vector of binary treatments}
\item{Y}{A vector of continuously valued outcome}
\item{distW}{A list containing relevant information needed to reproduce data sets}
\item{fnG0}{A list of lists containing relevant information needed to reproduce data sets}
\item{fnQ0}{A list of lists containing relevant information needed to reproduce data sets}
\item{distErrY}{A list containing relevant information needed to reproduce data sets}
}
\description{
Given an object output by \code{makeRandomData}, make another data set with the same
distribution.
}
