% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateFunctions.R
\name{sinUniParm}
\alias{sinUniParm}
\title{sinUniParm}
\usage{
sinUniParm(pLower = -1, pUpper = 1, ampLower = -1, ampUpper = 1, ...)
}
\description{
Draw random periodicity and amplitude for sinusoidal relationship univariate relationship.
See source code for distributions. 
 @export
}

