% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{uniformWCorParm}
\alias{uniformWCorParm}
\title{uniformWParm
Generate parameters for \code{uniformParm} (smaller lower = more correlation)}
\usage{
uniformWCorParm(lowerLower = -2, lowerUpper = 0, rangeLower = 0.25,
  rangeUpper = 5)
}
\description{
uniformWParm
Generate parameters for \code{uniformParm} (smaller lower = more correlation)
}
