% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateDistributions.R
\name{uniformWParm}
\alias{uniformWParm}
\title{uniformWParm}
\usage{
uniformWParm(lowerLower = -10, lowerUpper = 0, rangeLower = 0.5,
  rangeUpper = 10)
}
\description{
Generate parameters for \code{uniform}
}

